requirements.copy.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
azure-core==1.26.1
azure-identity==1.12.0
azure-storage-blob==12.14.1
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
defusedxml==0.7.1
dnspython==2.2.1
email-validator==1.3.0
et-xmlfile==1.1.0
fastapi==0.88.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
install==1.3.5
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.3
mypy-extensions==0.4.3
numpy==1.23.5
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
orjson==3.8.3
pandas==1.5.2
pathspec==0.10.2
platformdirs==2.6.0
portalocker==2.6.0
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
starlette-validation-uploadfile==0.1.1
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
